Disable production CI
There is an issue with the CI script which results in an unstyled
deployment. I have an issue to track fixing the problem, but in the mean
time we need to disable production CI deployment and just publish
manually to prevent accidental breaking of the live page.

Issue #4 Fix CI / CD