Add a download button
SVG is a very nice format but unfortunately cannot be downloaded
directly by a user as an image file. We unfortunately needed to convert
the svg to canvas, and then download the canvas in order for it to work.

This has a few side effects (images have to be directly embedded,
external fonts don't work, styles have to be embedded as well), but it
seems to work.

Issue #5